--- utils/exceptions.py
@@ -1,4 +1,0 @@
-
-class NotCorrectValue(Exception):
-    """when parse expanse/income/budget values """
-    pass
